[Rename] iterator -> interactor
