+++ app/src/main/java/com/withapp/coffeememo/di/search/recipe/use_case/GetAllRecipeUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.withapp.coffeememo.di.search.recipe.use_case
 
-import com.withapp.coffeememo.search.recipe.domain.iterator.GetAllRecipeIterator
+import com.withapp.coffeememo.search.recipe.domain.interacotr.GetAllRecipeInteractor
 import com.withapp.coffeememo.search.recipe.domain.use_case.GetAllRecipeUseCase
 import dagger.Binds
 import dagger.Module
@@ -13,6 +13,6 @@ abstract class GetAllRecipeUseCaseModule {
 
     @Binds
     abstract fun bindGetAllRecipeUseCase(
-        GetAllRecipeUseCaseImpl: GetAllRecipeIterator
+        GetAllRecipeUseCaseImpl: GetAllRecipeInteractor
     ): GetAllRecipeUseCase
 }
