+++ app/src/main/java/com/withapp/coffeememo/search/recipe/domain/interacotr/FilterRecipeIterator.kt
@@ -1,4 +1,4 @@
-package com.withapp.coffeememo.search.recipe.domain.iterator
+package com.withapp.coffeememo.search.recipe.domain.interacotr
 
 import com.withapp.coffeememo.domain.repository.RecipeRepository
 import com.withapp.coffeememo.search.recipe.data.mapper.SearchRecipeModelMapper
