+++ app/src/main/java/com/withapp/coffeememo/search/recipe/domain/interacotr/GetAllRecipeInteractor.kt
@@ -1,4 +1,4 @@
-package com.withapp.coffeememo.search.recipe.domain.iterator
+package com.withapp.coffeememo.search.recipe.domain.interacotr
 
 import com.withapp.coffeememo.domain.repository.RecipeRepository
 import com.withapp.coffeememo.search.recipe.data.mapper.SearchRecipeModelMapper
@@ -7,7 +7,7 @@ import com.withapp.coffeememo.search.recipe.domain.presenter.SearchRecipePresent
 import com.withapp.coffeememo.search.recipe.domain.use_case.GetAllRecipeUseCase
 import javax.inject.Inject
 
-class GetAllRecipeIterator @Inject constructor(
+class GetAllRecipeInteractor @Inject constructor(
     private val recipeRepo: RecipeRepository,
     private val mapper: SearchRecipeModelMapper,
     private val searchRecipePresenter: SearchRecipePresenter
